Move resource description utils
